refactor: do away with metrics unit tests

+++ das/checkpoint.go
@@ -50,10 +50,10 @@ func (c checkpoint) String() string {
 }
 
 // totalSampled returns the total amount of sampled headers
-func (c checkpoint) totalSampled() int {
+func (c checkpoint) totalSampled() uint64 {
 	totalInProgress := 0
 	for _, w := range c.Workers {
 		totalInProgress += int(w.To-w.From) + 1
 	}
-	return int(c.SampleFrom) - totalInProgress - len(c.Failed)
+	return uint64(int(c.SampleFrom) - totalInProgress - len(c.Failed))
 }
